Weird OP skipper coincidence
Believe it or not, but Crunchyroll chose to add a "SKIP INTRO" feature the day after I managed to make mine...
But their skipper is not working really well yet, mine is incomplete but it works 🙄.
And I'll continue to improve it.

I'll add a button to switch skippers later, I'm hidding theirs for now.